[sw/sca] OTBN vertical serial code restructuring
This commit restructures the serial code for OTBN vertical analysis.
In particular it splits up serial code for ecc256 keygen and ecc256
modinv into seperate source and header files, and combines them within a
single OTBN vertical serial file. This allows us to seperate the serial
code between the individual apps for more code clearance, but also to
have a single binary for all OTBN apps.
