+++ sw/device/sca/otbn_vertical/BUILD
@@ -6,87 +6,10 @@ load("//rules:opentitan.bzl", "opentitan_flash_binary")
 
 package(default_visibility = ["//visibility:public"])
 
-alias(
-    name = "aes_serial_aes_testutils",
-    actual = select({
-        "//sw/device:is_english_breakfast": "//sw/device:nothing",
-        "//conditions:default": "//sw/device/lib/testing:aes_testutils",
-    }),
-    visibility = ["//visibility:private"],
-)
-
-opentitan_flash_binary(
-    name = "aes_serial",
-    srcs = ["aes_serial.c"],
-    deps = [
-        "//hw/top_earlgrey/sw/autogen:top_earlgrey",
-        "//sw/device/lib/base:memory",
-        "//sw/device/lib/dif:aes",
-        "//sw/device/lib/runtime:log",
-        "//sw/device/lib/testing:aes_testutils",
-        "//sw/device/lib/testing/test_framework:check",
-        "//sw/device/lib/testing/test_framework:ottf_ld_silicon_creator_slot_a",
-        "//sw/device/lib/testing/test_framework:ottf_main",
-        "//sw/device/sca/lib:prng",
-        "//sw/device/sca/lib:sca",
-        "//sw/device/sca/lib:simple_serial",
-    ],
-)
-
-opentitan_flash_binary(
-    name = "kmac_serial",
-    srcs = ["kmac_serial.c"],
-    deps = [
-        "//hw/top_earlgrey/sw/autogen:top_earlgrey",
-        "//sw/device/lib/base:memory",
-        "//sw/device/lib/dif:kmac",
-        "//sw/device/lib/runtime:log",
-        "//sw/device/lib/testing/test_framework:ottf_ld_silicon_creator_slot_a",
-        "//sw/device/lib/testing/test_framework:ottf_main",
-        "//sw/device/sca/lib:prng",
-        "//sw/device/sca/lib:sca",
-        "//sw/device/sca/lib:simple_serial",
-    ],
-)
-
-opentitan_flash_binary(
-    name = "sha3_serial",
-    srcs = ["sha3_serial.c"],
-    deps = [
-        "//hw/top_earlgrey/sw/autogen:top_earlgrey",
-        "//sw/device/lib/base:memory",
-        "//sw/device/lib/dif:kmac",
-        "//sw/device/lib/runtime:log",
-        "//sw/device/lib/testing/test_framework:ottf_ld_silicon_creator_slot_a",
-        "//sw/device/lib/testing/test_framework:ottf_main",
-        "//sw/device/sca/lib:prng",
-        "//sw/device/sca/lib:sca",
-        "//sw/device/sca/lib:simple_serial",
-    ],
-)
-
-opentitan_flash_binary(
-    name = "ecc384_serial",
-    srcs = ["ecc384_serial.c"],
-    deps = [
-        "//hw/top_earlgrey/sw/autogen:top_earlgrey",
-        "//sw/device/lib/base:memory",
-        "//sw/device/lib/crypto/drivers:otbn",
-        "//sw/device/lib/runtime:ibex",
-        "//sw/device/lib/runtime:log",
-        "//sw/device/lib/testing:entropy_testutils",
-        "//sw/device/lib/testing/test_framework:ottf_ld_silicon_creator_slot_a",
-        "//sw/device/lib/testing/test_framework:ottf_main",
-        "//sw/device/sca/lib:prng",
-        "//sw/device/sca/lib:sca",
-        "//sw/device/sca/lib:simple_serial",
-        "//sw/otbn/crypto:p384_ecdsa_sca",
-    ],
-)
-
-opentitan_flash_binary(
-    name = "ecc256_sign_serial",
-    srcs = ["ecc256_sign_serial.c"],
+cc_library(
+    name = "ecc256_keygen_serial",
+    srcs = ["ecc256_keygen_serial.c"],
+    hdrs = ["ecc256_keygen_serial.h"],
     deps = [
         "//hw/top_earlgrey/sw/autogen:top_earlgrey",
         "//sw/device/lib/base:abs_mmio",
@@ -100,6 +23,49 @@ opentitan_flash_binary(
         "//sw/device/sca/lib:prng",
         "//sw/device/sca/lib:sca",
         "//sw/device/sca/lib:simple_serial",
-        "//sw/otbn/crypto:p256_ecdsa_sca",
+        "//sw/otbn/crypto:p256_key_from_seed_sca",
+    ],
+)
+
+cc_library(
+    name = "ecc256_modinv_serial",
+    srcs = ["ecc256_modinv_serial.c"],
+    hdrs = ["ecc256_modinv_serial.h"],
+    deps = [
+        "//hw/top_earlgrey/sw/autogen:top_earlgrey",
+        "//sw/device/lib/base:abs_mmio",
+        "//sw/device/lib/base:memory",
+        "//sw/device/lib/crypto/drivers:otbn",
+        "//sw/device/lib/runtime:ibex",
+        "//sw/device/lib/runtime:log",
+        "//sw/device/lib/testing:entropy_testutils",
+        "//sw/device/lib/testing/test_framework:ottf_ld_silicon_creator_slot_a",
+        "//sw/device/lib/testing/test_framework:ottf_main",
+        "//sw/device/sca/lib:prng",
+        "//sw/device/sca/lib:sca",
+        "//sw/device/sca/lib:simple_serial",
+        "//sw/otbn/crypto:p256_mod_inv_sca",
+    ],
+)
+
+opentitan_flash_binary(
+    name = "otbn_vertical_serial",
+    srcs = ["otbn_vertical_serial.c"],
+    deps = [
+        "//hw/top_earlgrey/sw/autogen:top_earlgrey",
+        "//sw/device/lib/base:abs_mmio",
+        "//sw/device/lib/base:memory",
+        "//sw/device/lib/crypto/drivers:otbn",
+        "//sw/device/lib/runtime:ibex",
+        "//sw/device/lib/runtime:log",
+        "//sw/device/lib/testing:entropy_testutils",
+        "//sw/device/lib/testing/test_framework:ottf_ld_silicon_creator_slot_a",
+        "//sw/device/lib/testing/test_framework:ottf_main",
+        "//sw/device/sca/lib:prng",
+        "//sw/device/sca/lib:sca",
+        "//sw/device/sca/lib:simple_serial",
+        "//sw/device/sca/otbn_vertical:ecc256_keygen_serial",
+        "//sw/device/sca/otbn_vertical:ecc256_modinv_serial",
+        "//sw/otbn/crypto:p256_mod_inv_sca",
     ],
 )
